Add `bc` to rpmbuilder packages and check for its presence
The `bc` command is now required for a successful run of the program. Therefore, it has been added to the list of packages installed by the rpmbuilder in `setup.sh`, and a check for its presence has also been added in the build file. If `bc` is not found, an error message will be displayed, and the script will abort.