Update distribution versions and improve build configuration
Updated versions of CentOS and Fedora in the defaults file. Added matrix.json and distro_versions.json files to structure information about distributions and their versions in a more manageable way. Also changed the GitHub Actions build configuration to use the new distro_versions.json file to set the matrix for the build.