Add support for CentOS 7 in crypt-keeper.sh
A conditional statement was added to the crypt-keeper.sh script to handle CentOS 7 specific scenario. If the DISTRO variable is "centos" and the RELEASE variable equals to 7, then the FROM_DISTRO and FROM_RELEASE_TAG are set to "getpagespeed/lts" and "el7" respectively. This change improves the overall compatibility of the script with different CentOS versions.